PB-554 Rename Controller APP to provider
