+++ app/config/settings_base.py
@@ -38,7 +38,7 @@ INSTALLED_APPS = [
     'django.contrib.contenttypes',
     'django.contrib.sessions',
     'django.contrib.messages',  #'django.contrib.staticfiles',
-    'controller'
+    'provider'
 ]
 
 MIDDLEWARE = [
