+++ app/provider/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class ControllerConfig(AppConfig):
+class ProviderConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'controller'
+    name = 'provider'
